+++ src/main/java/com/delivery/shop/shop/UpdateDeliveryAreaParam.java
@@ -1,0 +1,4 @@
+package com.delivery.shop.shop;
+
+public class UpdateDeliveryAreaRequest {
+}
